#104 fixed mining bugs (spam clicking, walking) #98  fixed teleport (and other skills) not showing exp given #96 fixed teleport coordinates #94 fixed dropping items when walking (its like RSC now) #93 fixed facing npcs that don't talk
