--- GameServer/src/org/moparscape/msc/gs/model/TileValue.java
@@ -1,11 +1,0 @@
-package org.moparscape.msc.gs.model;
-
-public class TileValue {
-	public int diagWallVal = 0;
-	public byte horizontalWallVal = 0;
-	public byte mapValue = 0;
-	public byte objectValue = 0;
-	public byte overlay = 0;
-	public byte verticalWallVal = 0;
-	public byte elevation = 0;
-}
