+++ GameServer/src/org/moparscape/msc/gs/phandler/client/DropHandler.java
@@ -4,6 +4,7 @@ import org.apache.mina.common.IoSession;
 import org.moparscape.msc.gs.Instance;
 import org.moparscape.msc.gs.connection.Packet;
 import org.moparscape.msc.gs.event.DelayedEvent;
+import org.moparscape.msc.gs.event.SingleEvent;
 import org.moparscape.msc.gs.model.InvItem;
 import org.moparscape.msc.gs.model.Item;
 import org.moparscape.msc.gs.model.Player;
@@ -43,6 +44,43 @@ public class DropHandler implements PacketHandler {
 			return;
 		}
 
+		// drop item after a path has finished
+		if(player.pathHandler != null && !player.pathHandler.finishedPath()) {
+			waitAndDrop(player, item);
+		} else {
+			drop(player, item);
+		}
+
+
+
+	}
+
+	public void waitAndDrop(final Player player,final InvItem item) {
+		Instance.getDelayedEventHandler().add(new SingleEvent(player, 500) {
+
+			@Override
+			public void action() {
+
+				if(owner.dropTickCount > 20) { // 10 seconds they are allowed to walk for. anything longer won't drop.
+					owner.dropTickCount = 0;
+					stop();
+				} else {
+					owner.dropTickCount++;
+					if(owner.pathHandler != null && !owner.pathHandler.finishedPath()) {
+						waitAndDrop(owner, item);
+					} else {
+						drop(owner, item);
+					}
+				}
+
+
+			}
+		});
+
+	}
+
+	public void drop(Player player, final InvItem item) {
+		player.getActionSender().sendMessage("Dropping!");
 		player.setStatus(Action.DROPPING_GITEM);
 		Instance.getDelayedEventHandler().add(new DelayedEvent(player, 500) {
 			public void run() {
@@ -57,8 +95,8 @@ public class DropHandler implements PacketHandler {
 				}
 				world.addEntryToSnapshots(new Activity(owner.getUsername(),
 						owner.getUsername() + " dropped ID: " + item.getID()
-								+ " amount: " + item.getAmount() + " at: "
-								+ owner.getX() + "/" + owner.getY()));
+						+ " amount: " + item.getAmount() + " at: "
+						+ owner.getX() + "/" + owner.getY()));
 
 				owner.getActionSender().sendSound("dropobject");
 				owner.getInventory().remove(item);
