+++ GameServer/src/org/moparscape/msc/gs/phandler/client/InvUseOnObject.java
@@ -20,7 +20,6 @@ import org.moparscape.msc.gs.external.ItemSmeltingDef;
 import org.moparscape.msc.gs.external.ItemSmithingDef;
 import org.moparscape.msc.gs.external.ItemWieldableDef;
 import org.moparscape.msc.gs.external.ReqOreDef;
-import org.moparscape.msc.gs.model.ActiveTile;
 import org.moparscape.msc.gs.model.Bubble;
 import org.moparscape.msc.gs.model.ChatMessage;
 import org.moparscape.msc.gs.model.GameObject;
@@ -29,6 +28,7 @@ import org.moparscape.msc.gs.model.MenuHandler;
 import org.moparscape.msc.gs.model.Npc;
 import org.moparscape.msc.gs.model.Player;
 import org.moparscape.msc.gs.model.World;
+import org.moparscape.msc.gs.model.landscape.ActiveTile;
 import org.moparscape.msc.gs.model.snapshot.Activity;
 import org.moparscape.msc.gs.phandler.PacketHandler;
 import org.moparscape.msc.gs.plugins.quests.Dorics;
