+++ GameServer/src/org/moparscape/msc/gs/phandler/client/ObjectAction.java
@@ -22,7 +22,6 @@ import org.moparscape.msc.gs.external.GameObjectDef;
 import org.moparscape.msc.gs.external.ObjectFishDef;
 import org.moparscape.msc.gs.external.ObjectFishingDef;
 import org.moparscape.msc.gs.external.ObjectWoodcuttingDef;
-import org.moparscape.msc.gs.model.ActiveTile;
 import org.moparscape.msc.gs.model.Bubble;
 import org.moparscape.msc.gs.model.ChatMessage;
 import org.moparscape.msc.gs.model.GameObject;
@@ -34,6 +33,7 @@ import org.moparscape.msc.gs.model.Path;
 import org.moparscape.msc.gs.model.Player;
 import org.moparscape.msc.gs.model.Point;
 import org.moparscape.msc.gs.model.World;
+import org.moparscape.msc.gs.model.landscape.ActiveTile;
 import org.moparscape.msc.gs.model.snapshot.Activity;
 import org.moparscape.msc.gs.phandler.PacketHandler;
 import org.moparscape.msc.gs.plugins.extras.Thieving;
@@ -402,7 +402,7 @@ public class ObjectAction implements PacketHandler {
 								if (object == null) {
 									return;
 								}
-								if (owner.getSpam()) {
+								if (owner.isPacketSpam()) {
 									return;
 								} else {
 									owner.setSpam(true);
@@ -420,7 +420,7 @@ public class ObjectAction implements PacketHandler {
 								if (object == null) {
 									return;
 								}
-								if (owner.getSpam()) {
+								if (owner.isPacketSpam()) {
 									return;
 								} else {
 									owner.setSpam(true);
@@ -498,7 +498,7 @@ public class ObjectAction implements PacketHandler {
 								if (object == null) {
 									return;
 								}
-								if (owner.getSpam())
+								if (owner.isPacketSpam())
 									return;
 
 								Thieving lock = new Thieving(owner, object);
