+++ GameServer/src/org/moparscape/msc/gs/phandler/client/TradeHandler.java
@@ -13,9 +13,9 @@ import org.moparscape.msc.gs.db.DBConnection;
 import org.moparscape.msc.gs.external.ItemDef;
 import org.moparscape.msc.gs.model.InvItem;
 import org.moparscape.msc.gs.model.Inventory;
-import org.moparscape.msc.gs.model.PathGenerator;
 import org.moparscape.msc.gs.model.Player;
 import org.moparscape.msc.gs.model.World;
+import org.moparscape.msc.gs.model.landscape.PathGenerator;
 import org.moparscape.msc.gs.model.snapshot.Activity;
 import org.moparscape.msc.gs.phandler.PacketHandler;
 import org.moparscape.msc.gs.tools.DataConversions;
