+++ GameServer/src/org/moparscape/msc/gs/phandler/client/WallObjectAction.java
@@ -10,13 +10,13 @@ import org.moparscape.msc.gs.event.ShortEvent;
 import org.moparscape.msc.gs.event.WalkToPointEvent;
 import org.moparscape.msc.gs.external.DoorDef;
 import org.moparscape.msc.gs.external.EntityHandler;
-import org.moparscape.msc.gs.model.ActiveTile;
 import org.moparscape.msc.gs.model.ChatMessage;
 import org.moparscape.msc.gs.model.GameObject;
 import org.moparscape.msc.gs.model.Npc;
 import org.moparscape.msc.gs.model.Player;
 import org.moparscape.msc.gs.model.Point;
 import org.moparscape.msc.gs.model.World;
+import org.moparscape.msc.gs.model.landscape.ActiveTile;
 import org.moparscape.msc.gs.model.snapshot.Activity;
 import org.moparscape.msc.gs.phandler.PacketHandler;
 import org.moparscape.msc.gs.plugins.extras.Thieving;
@@ -116,7 +116,7 @@ public class WallObjectAction implements PacketHandler {
 														"This feature is only avaliable on a members server");
 										return;
 									}
-									if (owner.getSpam()) {
+									if (owner.isPacketSpam()) {
 										return;
 									}
 									Thieving thiev = new Thieving(owner, object);
