+++ GameServer/src/org/moparscape/msc/gs/plugins/extras/Thieving.java
@@ -9,7 +9,6 @@ import org.moparscape.msc.gs.event.MiniEvent;
 import org.moparscape.msc.gs.event.ShortEvent;
 import org.moparscape.msc.gs.event.WalkToMobEvent;
 import org.moparscape.msc.gs.external.GameObjectDef;
-import org.moparscape.msc.gs.model.ActiveTile;
 import org.moparscape.msc.gs.model.Bubble;
 import org.moparscape.msc.gs.model.ChatMessage;
 import org.moparscape.msc.gs.model.GameObject;
@@ -18,6 +17,7 @@ import org.moparscape.msc.gs.model.Mob;
 import org.moparscape.msc.gs.model.Npc;
 import org.moparscape.msc.gs.model.Player;
 import org.moparscape.msc.gs.model.World;
+import org.moparscape.msc.gs.model.landscape.ActiveTile;
 import org.moparscape.msc.gs.states.Action;
 import org.moparscape.msc.gs.util.Logger;
 
@@ -305,7 +305,7 @@ public class Thieving {
 			Instance.getDelayedEventHandler().add(
 					new WalkToMobEvent(player, affectedMob, 1) {
 						public void arrived() {
-							if (owner.getSpam()) {
+							if (owner.isPacketSpam()) {
 								return;
 							} else {
